rename LoginModal source file to follow the convention
