+++ admin-dashboard/src/Layout.js
@@ -1,9 +1,7 @@
 import { Outlet } from "react-router-dom";
 import Navbar from './components/Navbar';
-import StyledFireBaseAuth from "./components/StyledFireBaseAuth";
 import './App.css';
-import { auth, authUiConfig } from './firebase-config'
-import LoginModal from "./components/modal";
+import LoginModal from "./components/LoginModal";
 
 const Layout = () => {
     return (
